# api/openapi.yaml
openapi: 3.0.3
info:
  title: x-ui
  version: "1.0.0"
  description: |
    Control-plane API for a tunnel panel: operator accounts, subscribers
    (clients), listeners (inbounds) with the TLS each terminates with, reported
    traffic, and the audit log.

    Every endpoint answers with the same envelope:

        { "success": true,  "msg": "", "obj": <payload> }
        { "success": false, "msg": "<reason>", "obj": null }

    `success` mirrors the status code; it is there so a client can branch on
    one field rather than on a status line it may not see through a proxy.

    The same endpoints are mounted twice, under two different authentication
    schemes. `{basePath}api` takes a session cookie and is what a browser uses;
    `{basePath}apiv2` takes a `Token` header and is what a script or a node
    uses. Only the cookie API enforces same-origin, because a cross-site page
    cannot set a custom header without a CORS preflight that is never answered.

    Only GET, POST and DELETE are used. An update is a POST to the record's own
    path rather than a PUT: a panel is reached through whatever sits in front of
    it — a corporate proxy, an embedded webview, a captive network's filter —
    and the three are the ones those are relied on to forward.
servers:
  - url: "http://localhost:8000"
    description: Panel — operators, authenticated
  - url: "http://localhost:8443"
    description: Subscriptions — subscribers, unauthenticated, its own listener
...
paths:
...
  /api/settings:
    get:
      tags: [settings]
      summary: Read the runtime settings
      description: |
        Only the options an operator may change, every value a string whatever
        it stands for. Anything needed to start the process comes from the
        configuration files. `/settings/startup` shows the running snapshot;
        `/settings/panel` reads and saves the editable startup options.

        | Key | Default | |
        |---|---|---|
        | `subUpdates` | `12` | Hours a client waits before fetching its subscription again |
        | `subEncode` | `true` | Base64 the `links` format |
        | `subShowInfo` | `false` | Name the remaining quota and expiry in the subscription |
        | `tgBotEnable` | `false` | Send notifications through the Telegram bot |
        | `tgBotToken` | | The bot's token, from @BotFather |
        | `tgBotChatIds` | | The chats notifications go to, comma separated; negative for a group |
        | `tgNotifySignIn` | `true` | Notify each operator sign-in |
        | `tgNotifyDeplete` | `true` | Notify subscribers taken offline for quota or expiry |
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      obj:
                        type: object
                        additionalProperties: { type: string }
    post:
      tags: [settings]
      summary: Write runtime settings
      description: |
        An unknown key is refused rather than stored: it would be written, read
        by nothing, and look to the operator like it had taken effect.

        The Telegram bot's settings are judged as they are written: a token
        has to be in the shape @BotFather hands one out in, a chat id is a
        number, and the bot is not switched on without a token and at least
        one chat — it would look set up and never send a thing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: { type: string }
      responses:
        "200":
          description: The settings after the write
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "400":
          $ref: "#/components/responses/Invalid"
  /api/settings/reset:
    post:
      tags: [settings]
      summary: Restore the runtime settings to their defaults
      description: |
        Named keys put back those alone, which is how one part of the settings
        page restores its own without taking the rest with it. No body, or no
        keys, puts back every one of them.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                keys:
                  type: array
                  items: { type: string }
                  example: [subUpdates, subEncode, subShowInfo]
      responses:
        "200":
          description: The settings after the reset
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      obj:
                        type: object
                        additionalProperties: { type: string }
        "400":
          $ref: "#/components/responses/Invalid"
...
  /api/settings/panel:
    get:
      tags: [settings]
      summary: Read saved Panel settings and the API startup snapshot
      responses:
        "200":
          description: Editable settings, file revision, environment overrides and restart status
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      obj: { $ref: "#/components/schemas/PanelSettingsState" }
    post:
      tags: [settings]
      summary: Save Panel settings to config.yaml for the next restart
      description: |
        Send the full typed form and the revision read before editing. Unknown
        fields and changes to environment-controlled fields are refused. The
        candidate is validated before atomically replacing the file, preserving
        other options, comments and file permissions. Restart the API and worker
        to apply it; this operation does not restart either service.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [revision, values]
              properties:
                revision: { type: string }
                values: { $ref: "#/components/schemas/PanelSettings" }
      responses:
        "200":
          description: Saved settings and restart status
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      obj: { $ref: "#/components/schemas/PanelSettingsState" }
        "400":
          $ref: "#/components/responses/Invalid"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/settings/panel/restart:
    post:
      tags: [settings]
      summary: Restart the panel to apply the saved configuration
      description: |
        Available for the systemd-managed installation. Saving is separate.
        The task survives API replacement, checks API/worker readiness and
        updates a locally managed agent connection if needed. The proxy core
        is kept running. Failed changes recover the previous Panel settings.
        Repeating a request for the same in-progress revision returns its task.
        Configuration saves are rejected while a task is in progress.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [revision]
              properties:
                revision: { type: string, minLength: 64, maxLength: 64 }
      responses:
        "202":
          description: Restart task accepted, not yet applied
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      obj: { $ref: "#/components/schemas/PanelRestartJob" }
        "400":
          $ref: "#/components/responses/Invalid"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/settings/panel/restart/{id}:
    get:
      tags: [settings]
      summary: Read the persisted Panel restart task result
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, pattern: "^[a-f0-9]{32}$" }
      responses:
        "200":
          description: Current restart progress or terminal result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      obj: { $ref: "#/components/schemas/PanelRestartJob" }
        "400":
          $ref: "#/components/responses/Invalid"
  /api/telegram/test:
    post:
      tags: [settings]
      summary: Send a test message through the Telegram bot
      description: |
        Sent to every chat with the bot as it is saved, switched on or not, so
        a token can be tried before anything depends on it. What Telegram said
        of a message it refused comes back as the reason.
      responses:
        "200":
          description: Every chat was sent the message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "400":
          description: No token or chat is saved, Telegram could not be reached, or it refused a message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /api/maintenance:
    get:
      tags: [settings]
      summary: Whether service is stopped on purpose
      responses:
        "200":
          description: The current state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
    post:
      tags: [settings]
      summary: Stop or resume service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enable]
              properties:
                enable: { type: boolean }
      responses:
        "200":
          description: The new state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: x-ui
      description: Set by /api/signin. Used under the `{basePath}api` prefix.
    token:
      type: apiKey
      in: header
      name: Token
      description: Minted by /api/tokens. Used under the `{basePath}apiv2` prefix.
...
  responses:
    Invalid:
      description: The request has to be fixed before it can succeed
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/Envelope"
              - type: object
                properties:
                  obj: { $ref: "#/components/schemas/ValidationErrors" }
...
    Conflict:
      description: Collides with something already stored
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
...
  schemas:
    Envelope:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        msg: { type: string }
        obj: { nullable: true }
    ValidationErrors:
      type: object
      properties:
        fields:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              reason: { type: string }
...
    PanelSettings:
      type: object
      additionalProperties: false
      description: Only the editable Panel options. Durations are whole seconds; no secrets are included.
      required: [listen, port, basePath, domain, keyFile, certFile, trustedProxies, maxAgeSeconds, statsRetentionSeconds, statsBucketSeconds, timeLocation, resetSpec, depleteSpec, cleanupSpec, logLevel]
      properties:
        listen: { type: string }
        port: { type: integer, minimum: 1, maximum: 65535 }
        basePath: { type: string, description: "Normalised to leading/trailing slashes. Segments use letters, numbers, - and _." }
        domain: { type: string }
        keyFile: { type: string }
        certFile: { type: string }
        trustedProxies:
          type: array
          items: { type: string }
        maxAgeSeconds: { type: integer, format: int64, minimum: 0, maximum: 9223372036 }
        statsRetentionSeconds: { type: integer, format: int64, minimum: 0, maximum: 9223372036 }
        statsBucketSeconds: { type: integer, format: int64, minimum: 1, maximum: 9223372036 }
        timeLocation: { type: string }
        resetSpec: { type: string, description: "Cron, descriptors, or empty/off to disable." }
        depleteSpec: { type: string, description: "Cron, descriptors, or empty/off to disable." }
        cleanupSpec: { type: string, description: "Cron, descriptors, or empty/off to disable." }
        logLevel: { type: string, enum: [debug, info, notice, warning, error, critical] }
    PanelSettingsState:
      type: object
      properties:
        saved: { $ref: "#/components/schemas/PanelSettings" }
        running: { $ref: "#/components/schemas/PanelSettings" }
        revision: { type: string, description: File revision to include in a save. }
        overrides:
          type: object
          additionalProperties: { type: string }
          description: Locked form field names mapped to their X_UI_* environment variable.
        restartRequired: { type: boolean, description: Saved settings require application or a restart task is still running. }
        restartSupported: { type: boolean }
        restartUnavailableReason: { type: string }
        restartJob: { $ref: "#/components/schemas/PanelRestartJob" }
    PanelRestartJob:
      type: object
      properties:
        id: { type: string, pattern: "^[a-f0-9]{32}$" }
        state: { type: string, enum: [queued, running, rolling_back, succeeded, rolled_back, failed] }
        revision: { type: string }
        actor: { type: string }
        requestedAt: { type: string, format: date-time }
        finishedAt: { type: string, format: date-time }
        values: { $ref: "#/components/schemas/PanelSettings" }
        previous: { $ref: "#/components/schemas/PanelSettings" }
        error: { type: string, description: Public application or recovery failure message. }
...
security:
  - session: []
  - token: []